Code_Homework_2/code/requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
apipkg==1.5
astroid==2.4.2
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.1
crayons==0.4.0
execnet==1.7.1
idna==2.10
iniconfig==1.0.1
isort==5.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-forked==1.3.0
pytest-xdist==2.1.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
webdriver-manager==3.2.2
wrapt==1.12.1
